+++ src/main/java/online/partyrun/partyrunbattleservice/domain/battle/entity/Battle.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.FieldDefaults;
 
-import online.partyrun.partyrunbattleservice.domain.runner.entuty.Runner;
+import online.partyrun.partyrunbattleservice.domain.runner.entity.Runner;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.Id;
