+++ src/main/java/online/partyrun/partyrunbattleservice/domain/runner/entity/RunnerStatus.java
@@ -1,0 +1,5 @@
+package online.partyrun.partyrunbattleservice.domain.runner.entity;
+
+public enum RunnerStatus {
+    READY, RUNNING, FINISHED
+}
