--- src/main/java/online/partyrun/partyrunbattleservice/domain/runner/entuty/RunnerStatus.java
@@ -1,5 +1,0 @@
-package online.partyrun.partyrunbattleservice.domain.runner.entuty;
-
-public enum RunnerStatus {
-    READY, RUNNING, FINISHED
-}
